Remove deprectated Makefile
The Makefile is deprecated. Use the new waf build system instead. The
ARCH option in the Makefile corresponds to the --os option in waf, the
TOOLCHAIN option corresponds to the --toolchain option and OUTDIR
corresponds to --libdir.

The equivalent of:
make OUTDIR=../bin/avr TOOLCHAIN=avr- ARCH=freertos clean all

is:
./waf configure --os=freertos --toolchain=avr- --libdir=../bin/avr build
install

Remember to add e.g. --with-can=at90can128 to enable CAN support on the
Atmel AT90CAN128. See ./waf configure --help for available options.