interfaces/can: move frame processing out of ISR
The CAN interface used to process incoming frames directly in the ISR.
This means the processor is blocked while a packet buffer is allocated,
the reassembly is performed, etc. This patch moves the processing to a
taks, and limits the ISR to simple queue up the frames.

The patch also includes a couple of cosmetic fixes to the code (use
static vars, use defines).