Fix csp_cmp so that it works for both set and get requests.
The disadvantage to this fix is that CMP ident requests now is the same
size as the reply. But the current macro method of defining message
lengths assumes that all requests will be empty, and all replies
contains data. However for route_set it's exactly the opposite. So this
should be fixed in the future.

The following command just worked:
cmp route_set 1 1000 2 2 KISS