interfaces: can: increase CAN RX thread priority
The CAN RX thread needs to have high priority, to ensure the frames are
moved from the RX queue to a proper packet as fast as possible.
Otherwise, the frames will take up unnecessary space in both the RX
queue and the buffer system.

Maybe we should add the thread priority and stack size as arguments to
csp_can_init?

Signed-off-by: Jeppe Ledet-Pedersen <jeppe@satlab.org>