Add a second type paramter allowing seperate Read/Wrtie Connections in the CommsControlBlock
This the CommsControlBlock can still be used in the same way as before
by setting the type paramter of WriteConnection and ReadConnection to
the same type and using one Arc<Mutex<..>>. The allows the read/write
threads to mutually lock eachother like before.

In the other case, Two seperate types can be used, two seperate
Arc<Mutex<..>> (Possibly dont need the mutex) allow the read/write
threads to operate independantly of eachother.